% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{assign_by_colnames}
\alias{assign_by_colnames}
\title{Loop through a list of data frames, and assign each to a name in a specified
environment, .GlobalEnv by default. The names are specified in a JSON file
which lives in the same directory as the csv file}
\usage{
assign_by_colnames(df_list, name_list, env = .GlobalEnv)
}
\arguments{
\item{df_list}{A list of data frames.}

\item{name_list}{A list of key-value pairs of the form list(list(col.names =
c("name1", "name2"), assigned_name = "out_name"), list(col.names =
c("name3"), assigned_name = "out_name2")). Such lists are generated from
JSON files (see example_name_list.JSON).}

\item{env}{An environment in which to store the names and data frames
specified by name_list. Defaults to .GlobalEnv.}
}
\value{
Alters the global environment, doesn't return any object.
}
\description{
Loop through a list of data frames, and assign each to a name in a specified
environment, .GlobalEnv by default. The names are specified in a JSON file
which lives in the same directory as the csv file
}

