% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{assign_by_colnames_0}
\alias{assign_by_colnames_0}
\title{Read the column names of a data frame.
If they match a target, then assign that data frame to a given
variable name.}
\usage{
assign_by_colnames_0(df, col.names, assigned_name, env = globalenv())
}
\arguments{
\item{df}{a data frame}

\item{col.names}{a character vector}

\item{assigned_name}{The name that df will be assigned in env if its
column names match col.names}
}
\value{
Alters the global environment, doesn't return any object.
}
\description{
Read the column names of a data frame.
If they match a target, then assign that data frame to a given
variable name.
}

