% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{assign_by_colnames_1}
\alias{assign_by_colnames_1}
\title{Given a data frame, loop through a list of (col.name, assigned_name) pairs,
If the column names match col.name, assign the data frame to the
corresponding assigned_name.}
\usage{
assign_by_colnames_1(df, name_list, ...)
}
\arguments{
\item{df}{A data frame.}

\item{name_list}{A list of key-value pairs of the form list(list(col.names =
c("name1", "name2"), assigned_name = "out_name"), list(col.names =
c("name3"), assigned_name = "out_name2")). Such lists are generated from
JSON files (see csv_test/name_list.JSON)}

\item{...}{additional parameters to pass to assign_by_colnames_0}
}
\value{
Alters the global environment, doesn't return any object.
}
\description{
Given a data frame, loop through a list of (col.name, assigned_name) pairs,
If the column names match col.name, assign the data frame to the
corresponding assigned_name.
}

