% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{assign_by_colnames_2}
\alias{assign_by_colnames_2}
\title{Given a list of data frames, and a list of (col.name, assigned_name) pairs,
loop through both and assign each data frame to the assigned_name that
matches its column names.}
\usage{
assign_by_colnames_2(df_list, name_list, ...)
}
\arguments{
\item{df_list}{A list of data frames}

\item{name_list}{A list of key-value pairs of the form list(list(col.names =
c("name1", "name2"), assigned_name = "out_name"), list(col.names =
c("name3"), assigned_name = "out_name2")). Such lists are generated from
JSON files (see csv_test/name_list.JSON)}
}
\value{
Alters the global environment, doesn't return any object.
}
\description{
Given a list of data frames, and a list of (col.name, assigned_name) pairs,
loop through both and assign each data frame to the assigned_name that
matches its column names.
}

