% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.r
\name{connect_to_postgres}
\alias{connect_to_postgres}
\title{Connect to Postgres via the RPostgreSQL package.}
\usage{
connect_to_postgres(auth_file_location = "~/.auth", ...)
}
\arguments{
\item{auth_file_location}{A character. Give the path of the directory where
your authenticate file is located. If left blank, R will search for a 
file named 'authenticate' in your current working directory. Set to NULL 
if you want to enter your credentials manually.}
}
\value{
A list containing a driver object, drv, and a connection object,
con.
}
\description{
Connect to Postgres via the RPostgreSQL package.
}
\examples{
connect_to_postgres(auth_file_location="~/authfiles")
}

