% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.r
\docType{data}
\name{gloo_colors}
\alias{gloo_colors}
\title{Gloo color palette}
\format{An object of class \code{character} of length 5.}
\usage{
gloo_colors
}
\description{
A named vector containing hex codes for the Gloo logo's colors
}
\keyword{datasets}

