% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.r
\docType{data}
\name{gloo_user_emails}
\alias{gloo_user_emails}
\title{Gloo internal emails}
\format{A character vector.}
\usage{
gloo_user_emails
}
\description{
A caracter vector containing all emails associated to Gloo employees'
'growing selves'.
}
\keyword{datasets}

