% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_and_name_csvs_from_directory}
\alias{load_and_name_csvs_from_directory}
\title{Load all csv files in a given directory, and assign them names in a given
environment (.GlobalEnv by default)
The names are assigned according to each data frame's column names. The
mapping from column names to assigned name is specified by a JSON file that
lives in the same directory as the csv files. See the
files in csv_test for an example of how to properly format said JSON file.}
\usage{
load_and_name_csvs_from_directory(path, ...)
}
\arguments{
\item{path}{The location of the directory to be loaded.}

\item{...}{Arguments to pass to assign_by_colnames_2}
}
\value{
Alters the global environment, doesn't return any object.
}
\description{
Load all csv files in a given directory, and assign them names in a given
environment (.GlobalEnv by default)
The names are assigned according to each data frame's column names. The
mapping from column names to assigned name is specified by a JSON file that
lives in the same directory as the csv files. See the
files in csv_test for an example of how to properly format said JSON file.
}

