% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_csv_from_directory}
\alias{load_csv_from_directory}
\title{Load all csv files from a given directory.}
\usage{
load_csv_from_directory(csv_path)
}
\arguments{
\item{csv_path}{A character specifying a directory where csv files are
located.}
}
\value{
A list of data frames, one for each csv file in csv_path.
}
\description{
Load all csv files from a given directory.
}

