% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_name_list_from_directory}
\alias{load_name_list_from_directory}
\title{Load a name list from a given directory.}
\usage{
load_name_list_from_directory(name_list_path)
}
\arguments{
\item{name_list_path}{A character vector specifying a directory where the
name list is located.}
}
\value{
a list of lists that can be used in the assign_by_colnames
functions.
}
\description{
Load a name list from a given directory.
}

