% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.r
\docType{data}
\name{load_test_2}
\alias{load_test_2}
\title{Dataset to test the load_data functions}
\format{data.frame}
\usage{
load_test_2
}
\description{
Dataset to test the load_data functions
}
\keyword{datasets}

