% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.r
\docType{data}
\name{query_pa_flash_cat}
\alias{query_pa_flash_cat}
\title{Platform action flash report category query.}
\format{A character vector.}
\usage{
query_pa_flash_cat
}
\description{
Creates a temporary table in Redshift that matches platform actions to their
Flash Report categories.
}
\keyword{datasets}

