% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.r
\docType{data}
\name{query_user_flash_cat}
\alias{query_user_flash_cat}
\title{User flash report category query.}
\format{A character vector.}
\usage{
query_user_flash_cat
}
\description{
Creates a temporary table in Redshift that matches users to their
Flash Report categories.
}
\keyword{datasets}

