% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_load_data.r
\name{run_look_list}
\alias{run_look_list}
\title{Run a list of looks and save the results in a list of
data frames.}
\usage{
run_look_list(look_list)
}
\arguments{
\item{look_list}{A list of (look_name, look_id) pairs.
See gloograph::test_looks for an example of how to format this list.}
}
\value{
A named list of data frames, one for each (look_name, look_id) pair.
}
\description{
Run a list of looks and save the results in a list of
data frames.
}

