% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_load_data.r
\name{run_query_list}
\alias{run_query_list}
\title{Run a list of queries specified by a list, and save the results in a list of
data frames.}
\usage{
run_query_list(query_list, connection = con)
}
\arguments{
\item{query_list}{A list of (query_name, query) pairs.
See gloograph::test_queries for an example of how to format this list.}

\item{connection}{The production database's PostgreSQL connection object.}
}
\value{
A named list of data frames, one for each (query_name, query) pair.
}
\description{
Run a list of queries specified by a list, and save the results in a list of
data frames.
}

