% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.r
\docType{data}
\name{test_looks}
\alias{test_looks}
\title{Looks to test the load_looks functions.}
\format{list of strings}
\usage{
test_looks
}
\description{
Looks to test the load_looks functions.
}
\keyword{datasets}

