% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.r
\docType{data}
\name{test_queries}
\alias{test_queries}
\title{Queries to test the load_queries functions.}
\format{list of strings}
\usage{
test_queries
}
\description{
Queries to test the load_queries functions.
}
\keyword{datasets}

