% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.r
\name{tufte_grouped_bar_chart}
\alias{tufte_grouped_bar_chart}
\title{Create a Tufte-style grouped bar chart}
\usage{
tufte_grouped_bar_chart(df, ybreaks = 0, x_var, y_var, group,
  theme_tufte_opts = list(base_size = 36, ticks = T),
  geom_bar_opts = list(postition = "dodge", stat = "identity", width = 0.25),
  ...)
}
\arguments{
\item{df}{A data frame containing the data to be plotted}

\item{ybreaks}{A vector determining where the y-axis breaks are.}

\item{x_var}{Character that names the column of df to be used for
the x variable.}

\item{y_var}{Character that names the column of df to be used for
the y variable.}

\item{group}{Character that determines which column of df to group by.}

\item{theme_tufte_opts}{Arguments to be passed to ggthemes::theme_tufte}

\item{geom_bar_opts}{Arguments to be passed to ggplot2::geom_bar}

\item{...}{Arguments to pass to scale_y_continuous}
}
\value{
A ggplot object
}
\description{
Create a Tufte-style grouped bar chart
}

