% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{assign_by_colnames_1}
\alias{assign_by_colnames_1}
\title{Read the column names of a data frame.
If they match a target, then assign that data frame to a given
variable name.}
\usage{
assign_by_colnames_1(df, name_list, ...)
}
\arguments{
\item{df}{a data frame}

\item{name_list}{A list of key-value pairs of the form list(list(col.names =
c("name1", "name2"), assigned_name = "out_name"), list(col.names =
c("name3"), assigned_name = "out_name2")). Such lists are generated from
JSON files (see example_name_list.JSON).}

\item{col.names}{a character vector}

\item{assigned_name}{The name that df will be assigned in env if its
column names match col.names}

\item{df}{A data frame.}
}
\value{
Alters the global environment, doesn't return any object.
Given a data frame, loop through a list of (col.name, assigned_name) pairs,
If the column names match col.name, assign the data frame to the
corresponding assigned_name.

Alters the global environment, doesn't return any object.
}
\description{
Read the column names of a data frame.
If they match a target, then assign that data frame to a given
variable name.
}

