% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.r
\name{connect_to_lookr}
\alias{connect_to_lookr}
\title{Connect to Looker via the LookR package.}
\usage{
connect_to_lookr(auth_file_location = "~/.auth", ...)
}
\arguments{
\item{auth_file_location}{A character. Give the path of the directory where
your authenticate file is located. If left blank, R will search for a 
file named 'authenticate' in your current working directory. Set to NULL 
if you want to enter your credentials manually.}
}
\value{
Connects your R session to the Looker API. No objects are returned.
}
\description{
Connect to Looker via the LookR package.
}
\examples{
connect_to_looker(auth_file_location="~/authfiles")
}

