% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_and_name_csvs_from_directory}
\alias{load_and_name_csvs_from_directory}
\title{Load all csv files in a given directory, and assign them names in a given
environment (.GlobalEnv by defaul)}
\usage{
load_and_name_csvs_from_directory(path, ...)
}
\arguments{
\item{path}{The location of the directory to be loaded.}
}
\value{
Alters the global environment, doesn't return any object.
}
\description{
Load all csv files in a given directory, and assign them names in a given
environment (.GlobalEnv by defaul)
}

