% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.r
\name{tufte_bar_chart}
\alias{tufte_bar_chart}
\title{Create a Tufte-style bar chart}
\usage{
tufte_bar_chart(df, ybreaks = 0, x_var, y_var,
  theme_tufte_opts = list(base_size = 36, ticks = T), ...)
}
\arguments{
\item{df}{A data frame containing the data to be plotted}

\item{ybreaks}{A vector determining where the y-axis breaks are.}

\item{x_var, }{y_var Characters that name the columns of df to be used for
the x and y variables, respectively.}

\item{theme_tufte_opts}{Arguments to be passed to ggthemes::theme_tufte}
}
\value{
A ggplot object
}
\description{
Create a Tufte-style bar chart
}

